micronucleus: fix compatibility with v1 boards
The code here was patching the user reset vector on both v1 and v2
protocols; even though the device takes it as its responsibility to
handle patching on v1. This resulted in a case where the device was
constantly jumping to the bootloader, as it interpreted our patched
"jump to bootloader" as the user program's reset vector.